% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginApp.R
\name{loginAppServer}
\alias{loginAppServer}
\title{Server For Shiny Login}
\usage{
loginAppServer(id, app_body = NULL, app_sidebar = NULL, config = NULL)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Server For Shiny Login
}
\examples{
#' ## Only run this example in interactive R sessions
if (interactive()) {
  ui <- loginAppUI('app')

  server <- function(input, output, session) {
  output$neil <- loginAppServer('app')
 }

 shinyApp(ui = ui, server = server)
}
}
