% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginApp.R
\name{loginAppUI}
\alias{loginAppUI}
\title{UI For Shiny Login}
\usage{
loginAppUI(id)
}
\arguments{
\item{id}{ID to read the value from}
}
\value{

}
\description{
UI For Shiny Login
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  ui <- loginAppUI('app')

  server <- function(input, output, session) {
  output$neil <- loginAppServer('app')
 }

 shinyApp(ui = ui, server = server)
}
}
