% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{logindb_create}
\alias{logindb_create}
\title{Create the login user database}
\usage{
logindb_create()
}
\arguments{
\item{con}{A database connection}
}
\value{

}
\description{
Create the login user database
}
