% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{user_add}
\alias{user_add}
\title{Add a user to the application}
\usage{
user_add(user_email = NULL, username = NULL, password = NULL)
}
\arguments{
\item{username}{username}

\item{password}{Hashed password}

\item{con}{User database connection}

\item{email}{email address}
}
\value{
TRUE
}
\description{
Add a user to the application
}
\examples{
con <- dbConnect(RSQLite::SQLite(), ":memory:")

login_db_create(con)

user_add(
  email = "user(at)example.com",
  username = "another_user",
  password = "ahashedpassword"
)

}
