% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{user_exists}
\alias{user_exists}
\title{Check Whether A User Exists in the Login Database}
\usage{
user_exists(email_new = "", username_new = "")
}
\arguments{
\item{con}{A database connection}

\item{email}{User email}

\item{username}{Username}
}
\value{
TRUE if the checked parameter exists, otherwise FALSE
}
\description{
Check Whether A User Exists in the Login Database
}
